@@ -1,6 +1,6 @@
 module github.com/jinmukeji/plat-pkg/v2
 
-go 1.14
+go 1.16
 
 replace (
 	// FIXME: 由于 etcd 与 gRPC 的兼容问题，暂时使用定制的 etcd 版本
@@ -21,7 +21,7 @@ require (
 	github.com/gin-gonic/gin v1.6.3
 	github.com/go-sql-driver/mysql v1.5.0
 	github.com/gobwas/glob v0.2.3
-	github.com/jinmukeji/go-pkg/v2 v2.5.2
+	github.com/jinmukeji/go-pkg/v2 v2.6.0
 	github.com/jinzhu/gorm v1.9.16
 	github.com/micro/cli/v2 v2.1.2
 	github.com/micro/go-micro/v2 v2.9.1
@@ -31,7 +31,7 @@ require (
 	github.com/micro/go-plugins/wrapper/service/v2 v2.9.1
 	github.com/micro/micro/v2 v2.9.3
 	github.com/rs/xid v1.2.1
-	github.com/sirupsen/logrus v1.7.0
+	github.com/sirupsen/logrus v1.8.1
 	github.com/smallstep/cli v0.15.0
 	github.com/stretchr/testify v1.7.0
 	google.golang.org/grpc v1.26.0
